refactor: 서비스를 application 패키지로 이동
